LineGroupClass_testing
I made a line group class, using 7 parameters, startX, startY, endX, endY, angle of line, number of lines, and color.
The class has a draw function, using a forloop to control the size of the parallelogram. Also, the lines in each group are made parallel by giving a line spacing of 10. The distance between each starting points and between end points of the lines are calculated using the angle and spacing.
The line groups stored into an array named as groups.
In the main draw function, a for loop is used to draw each line group.